XIVY-11352 Fix resource leak in LogBean >> R10
